RESKC-445: STE When you click on the Proposition Inquiry link inside a Rule Inquiry Screen.
As a Functional Administrator I need to be able to review the full logic included in a rule when designing Questionnaires and Agendas.

Steps to Recreate:
1 Create a New Agenda and Add a Rule
2 Click on the Search Link for Copy a Rule
3 Search for a Rule and click on the inquiry link for that rule
4 Click on a Proposition ID link in the rule
Result:
When you do this you get a STE

Acceptance Criteria
Given that I have clicked on a Proposition ID link in a Rule Inquiry Screen, the full logic for the proposition should appear on the inquiry screen that opens. (this would be consistent with the results in the Rule section of the Agenda document)